requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django~=3.2.6
django-admin-sortable2~=1.0
django-cache-url==3.2.3
django-tinymce==3.3.0
environs~=9.3.2
marshmallow==3.13.0
Pillow==8.3.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
requests~=2.26.0